refactor split init
